unseparating components
